Update schema to change total_amount type to DECIMAL and refactor checkout view for improved structure and success alert handling

+++ Database/shema.sql
@@ -69,7 +69,7 @@
 
     user_id INT NOT NULL,
 
-    total_amount INT NOT NULL,
+    total_amount DECIMAL(10, 2) NOT NULL,
 
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
